REFACTOR : HttpResponse 속의 static 메소드를 HttpResponseUtils 클래스로 리팩토링
