+++ src/main/java/Controller/StaticFileController.java
@@ -3,6 +3,7 @@ package Controller;
 import http.HttpRequest;
 import http.HttpResponse;
 import http.HttpStatus;
+import util.HttpResponseUtils;
 
 import java.io.IOException;
 
@@ -11,8 +12,8 @@ public class StaticFileController implements Controller{
     @Override
     public HttpResponse makeResponse(HttpRequest httpRequest) throws IOException {
         String contentType = httpRequest.getContentType();
-        String filePath = HttpResponse.makeFilePath(contentType);
-        byte[] responseBody = HttpResponse.makeBody(httpRequest, filePath);
+        String filePath = HttpResponseUtils.makeFilePath(contentType);
+        byte[] responseBody = HttpResponseUtils.makeBody(httpRequest, filePath);
         return new HttpResponse(HttpStatus.OK, responseBody, contentType);
     }
 }
